fix: fix SendActivity flikering issue by turning async Lg api call to sync (#2173)
* move `useWindowDimensions` to separated file

* move `useLgTemplate` under hooks folder

* use `lgFiles` in shell store to compute lg text

* move useLgTemplate to 'hooks' folder
