+++ Composer/packages/extensions/visual-designer/src/store/NodeRendererContext.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 import React from 'react';
+import { LGTemplate } from 'botbuilder-lg';
 import { ShellApi } from '@bfc/shared';
 
 type ShellApiFuncs =
@@ -12,11 +13,12 @@ type ShellApiFuncs =
   | 'updateLgTemplate'
   | 'removeLuIntent';
 
-interface NodeRendererContextValue extends Pick<ShellApi, ShellApiFuncs> {
+export interface NodeRendererContextValue extends Pick<ShellApi, ShellApiFuncs> {
   focusedId?: string;
   focusedEvent?: string;
   focusedTab?: string;
   clipboardActions: any[];
+  getLgTemplateSync: (lgTemplateName: string) => LGTemplate | undefined;
 }
 
 export const NodeRendererContext = React.createContext<NodeRendererContextValue>({
@@ -24,6 +26,7 @@ export const NodeRendererContext = React.createContext<NodeRendererContextValue>
   focusedEvent: '',
   focusedTab: '',
   clipboardActions: [],
+  getLgTemplateSync: () => undefined,
   getLgTemplates: () => Promise.resolve([]),
   copyLgTemplate: () => Promise.resolve(''),
   removeLgTemplate: () => Promise.resolve(),
